feat: move arena allocator to size_t

+++ software/luna/include/arena_allocator.h
@@ -6,11 +6,11 @@
 
 struct ArenaAllocator {
   void *arena;
-  uint16_t length;
-  uint16_t capacity;
+  size_t length;
+  size_t capacity;
 };
 
-struct ArenaAllocator arena_make(void *arena, uint16_t capacity);
+struct ArenaAllocator arena_make(void *arena, size_t capacity);
 void *arena_alloc(struct ArenaAllocator *allocator, size_t size);
 
 #endif
